Backend/requirements.txt
flask-restful
waitress
Flask-SQLAlchemy==2.5.1
Flask==1.1.2
python-dotenv
flask-cors
Flask-Script
Flask-Migrate
psycopg2-binary
requests









